Move LineLength and LineWidth out of settings
