+++ crates/ruff/src/line_width.rs
