+++ crates/ruff/src/rules/flake8_simplify/rules/ast_with.rs
@@ -9,8 +9,8 @@ use ruff_python_ast::helpers::{first_colon_range, has_comments_in};
 use ruff_python_ast::newlines::StrExt;
 
 use crate::checkers::ast::Checker;
+use crate::line_width::LineWidth;
 use crate::registry::AsRule;
-use crate::settings::line_width::LineWidth;
 
 use super::fix_with;
