+++ crates/ruff/src/settings/mod.rs
@@ -25,13 +25,12 @@ use crate::settings::configuration::Configuration;
 use crate::settings::types::{FilePatternSet, PerFileIgnore, PythonVersion, SerializationFormat};
 use crate::warn_user_once_by_id;
 
-use self::line_width::{LineLength, TabSize};
 use self::rule_table::RuleTable;
+use super::line_width::{LineLength, TabSize};
 
 pub mod configuration;
 pub mod defaults;
 pub mod flags;
-pub mod line_width;
 pub mod options;
 pub mod options_base;
 pub mod pyproject;
